requirements.txt
frappe>=14.0.0
pytesseract>=0.3.8
opencv-python>=4.5.0
Pillow>=8.0.0  # Add this since it's in setup.py